requirements.txt
streamlit>=1.32.0
anthropic>=0.40.0
mcp>=1.0.0
pymupdf4llm>=0.0.17
pymupdf>=1.24.0
piper-tts>=1.4.0
piper-phonemize>=1.1.0
lameenc>=1.7.0
